fix(config): resolve the auto main-segment size in the config layer (#727)
CalculateMainSegmentSize() returned 0 as an "auto" sentinel and left the
real sizing to IpcManager::ServerInitShm(). But GetMemorySegmentSize(
kMainSegment) forwards straight to it, so the public accessor started
reporting a 0-byte main segment -- cr_autogen_coverage_tests (and its
force_net variant) failed 'size > 0' on all 17 build/boost/icx jobs.

Resolve the auto case where the value is read instead of where the
segment is created, so no caller ever sees the sentinel:

- ctp::SystemInfo::GetProcessMemoryBudget() is the new home for the
  cgroup-aware budget; ReadCgroupMemoryLimit() moves there from
  ipc_manager.cc, which was always the wrong layer for /sys/fs/cgroup
  parsing and is now needed by two callers.
- CalculateMainSegmentSize() computes the auto default itself:
  min(1 GiB, max(64 MiB, budget/4)), flat 1 GiB when the budget is
  unknown. Never returns 0.
- ServerInitShm() keeps only the half-budget clamp, now applied
  uniformly -- a no-op for the auto value, which is already <= budget/4.

Same number reaches shm_init, just computed one layer up.
